refactor service and repo layer
