+++ internal/server/user/service/create_user.go
@@ -1,0 +1,18 @@
+package service
+
+import (
+	"context"
+
+	"github.com/go-park-mail-ru/2025_1_sigmaScript/internal/server/models"
+	"github.com/rs/zerolog/log"
+)
+
+func (s *UserService) CreateUser(ctx context.Context, user *models.User) error {
+	err := s.repo.CreateUser(ctx, user)
+	if err != nil {
+		log.Error().Err(err).Msg(err.Error())
+		return err
+	}
+
+	return nil
+}
